refactor: Env module
