+++ code/sys.driver/driver-vitepress/src/-tmpl/mod.ts
@@ -1,2 +1,1 @@
 export { Tmpl } from './m.Tmpl.ts';
-export { ensureFiles } from './u.ensureFiles.ts';
