+++ code/sys.driver/driver-vitepress/src/m.VitePress/common.ts
@@ -1,4 +1,5 @@
 export { ViteLog } from '@sys/driver-vite/log';
-export { Tmpl, ensureFiles } from '../-tmpl/mod.ts';
+export { Tmpl } from '../-tmpl/mod.ts';
+export { Env } from '../m.Env/mod.ts';
 
 export * from '../common.ts';
