--- code/sys.driver/driver-vitepress/src/m.VitePress/m.Env.ts
@@ -1,9 +1,0 @@
-import type { t } from './common.ts';
-import { init } from './u.initEnv.ts';
-
-/**
- * Helpers for establishing and updating the project environment.
- */
-export const Env: t.VitePressEnvLib = {
-  init,
-};
