+++ code/sys.driver/driver-vitepress/src/m.VitePress/m.VitePress.ts
@@ -1,5 +1,6 @@
 import type { t } from './common.ts';
-import { Env } from './m.Env.ts';
+
+import { Env } from '../m.Env/mod.ts';
 import { build } from './u.cmd.build.ts';
 import { dev } from './u.cmd.dev.ts';
