+++ code/sys.driver/driver-vitepress/src/m.VitePress/u.cmd.build.ts
@@ -1,5 +1,4 @@
-import { type t, Cmd, Cli, Fs, Pkg, Time, c } from './common.ts';
-import { Env } from './m.Env.ts';
+import { type t, c, Cli, Cmd, Env, Fs, Pkg, Time } from './common.ts';
 import { Log } from './u.log.ts';
 
 type B = t.VitePressLib['build'];
