--- code/sys.driver/driver-vitepress/src/m.VitePress/u.initEnv.ts
@@ -1,12 +1,0 @@
-import { type t, c, ensureFiles } from './common.ts';
-
-type F = t.VitePressEnvLib['init'];
-
-export const init: F = async (args = {}) => {
-  const { inDir = '', srcDir, force = false, silent = false, filter } = args;
-  const files = await ensureFiles({ inDir, srcDir, force, filter });
-  if (!silent) {
-    console.info(c.green('Update Environment'));
-    files.table.render();
-  }
-};
