+++ code/sys.driver/driver-vitepress/src/types.ts
@@ -3,5 +3,5 @@
  * Module types.
  */
 export type * from './-cmd/t.ts';
-export type * from './m.Config/t.ts';
+export type * from './m.Env/t.ts';
 export type * from './m.VitePress/t.ts';
